jbot/requirements.txt
qrcode==7.3.1
Telethon==1.24.0
requests==2.27.1
Pillow==9.0.0
httpx==0.21.3
python-socks==2.0.1
async_timeout==4.0.2
prettytable==3.0.0
speedtest-cli>=2.1.3
pyzbar==0.1.9